+++ docs/astro.config.ts
@@ -9,7 +9,7 @@ import { svgSprite } from './src/integrations/svgSprite';
 
 // https://astro.build/config
 export default defineConfig({
-  site: process.env.NODE_ENV === 'production' ? 'https://vee-validate.logaretm.com/' : 'http://localhost:3000/',
+  site: process.env.NODE_ENV === 'production' ? 'https://vee-validate.logaretm.com/' : 'http://localhost:4321/',
   trailingSlash: 'always',
   base: '/v4',
   integrations: [